requirements.txt
# Python dependencies
streamlit==1.40.1
web3==7.2.0
python-dotenv==1.0.1
py-solc-x==2.0.3
eth-account==0.13.4
pytest==8.3.3

